Weights are now doubles

+++ Source/Graph.hpp
@@ -7,19 +7,19 @@ using namespace std;
 class Graph
 {
 	public:
-		vector<vector<int>> matrix;
+		vector<vector<double>> matrix;
 		
 		Graph(int vert_count, double density);
 
 		vector<int> neighbors(int x);
 
-		inline void add_edge(int x, int y, int weight);
+		inline void add_edge(int x, int y, double weight);
 
 		inline bool adjacent(int x, int y);
 
 		inline void delete_edge(int x, int y);
 
-		inline int edge_weight(int x, int y) { return matrix[x][y]; };
+		inline double edge_weight(int x, int y) { return matrix[x][y]; };
 
 		void print_Graph();
